Implement methods directly on the client
