+++ Cargo.toml
@@ -13,6 +13,7 @@ categories = ["api-bindings", "wasm"]
 [dependencies]
 http = "1.3"
 lemmy_api_common = { git = "https://github.com/LemmyNet/lemmy", branch = "main" }
+reqwest = { version = "0.12", features = ["json"], default-features = false }
 serde = "1.0"
 serde_json = "1.0"
 
@@ -21,14 +22,3 @@ default = ["default-tls"]
 default-tls = ["reqwest/default-tls"]
 native-tls = ["reqwest/native-tls"]
 rustls-tls = ["reqwest/rustls-tls"]
-
-[target.'cfg(target_family = "wasm")'.dependencies]
-gloo-net = { version = "0.6", features = [
-  "http",
-  "json",
-], default-features = false }
-serde_urlencoded = "0.7"
-wasm-bindgen = "0.2"
-
-[target.'cfg(not(target_family = "wasm"))'.dependencies]
-reqwest = { version = "0.12", features = ["json"], default-features = false }
